add old tests
